+++ src/onegov.election_day/onegov/election_day/views/manage/subscribers.py
@@ -18,7 +18,7 @@ def view_subscribers(self, request):
 
     return {
         'layout': ManageSubscribersLayout(self, request),
-        'title': _("Manage"),
+        'title': _("Subscribers"),
         'count': self.query().count(),
         'subscribers': self.batch,
         'term': self.term